feat(Introduction): add introduction for default course loading and completion flag (#44)
This commit adds a new feature to the SmartsGrades app that introduces
an interactive introduction where the user is asked about their
university and major. Based on this information, the app loads default
courses relevant to the user's major. Additionally, once the user
completes the introduction, a flag is set to indicate that it has been
done, preventing the user from accessing it again.

---------

Signed-off-by: Valentin Sickert <17144397+Lapotor@users.noreply.github.com>