feat: update MacBook landing page with new dashboard look (#72)
This commit introduces a visual update to the SmartsGrades app's landing
page. The MacBook image on the landing page has been updated to reflect
the new look of the dashboard.

By incorporating the new dashboard look into the MacBook image on the
landing page, users can get a preview of the updated interface and have
a better understanding of the app's visual style and features.